changes to std libs
